+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in covermymeds_api.gemspec
+# Specify your gem's dependencies in cover_my_meds.gemspec
 gemspec
