+++ lib/cover_my_meds/client.rb
