+++ lib/cover_my_meds/client/credentials.rb
