+++ lib/cover_my_meds/client/drugs.rb
