+++ lib/cover_my_meds/client/indicators.rb
