+++ lib/cover_my_meds/client/pharmacies.rb
