+++ lib/cover_my_meds/error.rb
