+++ lib/cover_my_meds/meta.rb
