+++ lib/cover_my_meds/meta/host_and_path.rb
